+++ R/fitLinear.R
@@ -17,9 +17,9 @@
 #' than two arms. By default, it is not specified,
 #' all data will be used to fit the model. More than one condition can be
 #' specified in \code{...}, e.g.,
-#' \code{fitLinear(cfb ~ arm, 'pbo', data, arm \%in\% c('pbo', 'low dose'), cfb > 0.5)},
+#' \code{fitLinear(cfb ~ arm, 'pbo', data, 'greater', arm \%in\% c('pbo', 'low dose'), cfb > 0.5)},
 #' which is equivalent to:
-#' \code{fitLinear(cfb ~ arm, 'pbo', data, arm \%in\% c('pbo', 'low dose') & cfb > 0.5)}.
+#' \code{fitLinear(cfb ~ arm, 'pbo', data, 'greater', arm \%in\% c('pbo', 'low dose') & cfb > 0.5)}.
 #' Note that if more than one treatment arm are present in the data after
 #' applying filter in \code{...}, models are fitted for placebo verse
 #' each of the treatment arms.
@@ -28,12 +28,13 @@
 #' \describe{
 #' \item{\code{arm}}{name of the treatment arm. }
 #' \item{\code{placebo}}{name of the placebo arm. }
+#' \item{\code{estimate}}{estimate depending on \code{scale}. }
 #' \item{\code{p}}{one-sided p-value for between-arm difference (treated vs placebo). }
 #' \item{\code{info}}{sample size used in model with \code{NA} being removed. }
 #' \item{\code{z}}{z statistics of between-arm difference (treated vs placebo). }
 #' }
 #'
-#' @importFrom stats glm
+#' @importFrom stats glm pt
 #' @importFrom emmeans emmeans contrast
 #' @export
 #'
@@ -93,20 +94,22 @@ fitLinear <- function(formula, placebo, data, alternative, ...) {
     # Fit the linear regression model
     fit <- glm(formula, data = sub_data, family = 'gaussian')
 
-    ref_grid <- emmeans(fit, ~ arm)
-    cont <- contrast(
-      ref_grid,
-      method = list('trt_vs_pbo' = setNames(c(-1, 1), c(placebo, trt_arm)))
-    ) %>%
+    res <- fit %>%
+      emmeans(~ arm) %>%
+      contrast(method =
+                 list('trt_vs_pbo' = setNames(c(-1, 1), c(placebo, trt_arm)))
+               ) %>%
       summary()
 
-    z <- cont$t.ratio
-    df <- cont$df
+    estimate <- res$estimate
+    z <- res$t.ratio
+    df <- res$df
     p <- ifelse(alternative == 'greater', 1 - pt(z, df = df), pt(z, df = df))
 
     info <- fit$df.residual + fit$rank
 
     ret <- rbind(ret, data.frame(arm = trt_arm, placebo = placebo,
+                                 estimate = estimate,
                                  p = p, info = info, z = z
                                  )
                  )